+++ src/OpenSimCreator/Graphics/ModelRendererParams.cpp
@@ -8,8 +8,8 @@
 #include <oscar/Graphics/Scene/SceneRendererParams.h>
 #include <oscar/Maths/PolarPerspectiveCamera.h>
 #include <oscar/Platform/AppSettings.h>
-#include <oscar/Platform/AppSettingValue.h>
 #include <oscar/Utils/Algorithms.h>
+#include <oscar/Variant/Variant.h>
 
 #include <string>
 #include <string_view>
@@ -19,13 +19,13 @@ using namespace osc;
 
 namespace
 {
-    std::unordered_map<std::string, AppSettingValue> ToValues(
+    std::unordered_map<std::string, Variant> ToValues(
         std::string_view prefix,
         const ModelRendererParams& params)
     {
-        std::unordered_map<std::string, AppSettingValue> rv;
+        std::unordered_map<std::string, Variant> rv;
         std::string subPrefix;
-        const auto callback = [&subPrefix, &rv](std::string_view subkey, AppSettingValue value)
+        const auto callback = [&subPrefix, &rv](std::string_view subkey, Variant value)
         {
             rv.insert_or_assign(subPrefix + std::string{subkey}, std::move(value));
         };
@@ -36,8 +36,8 @@ namespace
         params.overlayOptions.forEachOptionAsAppSettingValue(callback);
         subPrefix = std::string{prefix} + std::string{"graphics/"};
         params.renderingOptions.forEachOptionAsAppSettingValue(callback);
-        rv.insert_or_assign(std::string{prefix} + "light_color", AppSettingValue{params.lightColor});
-        rv.insert_or_assign(std::string{prefix} + "background_color", AppSettingValue{params.backgroundColor});
+        rv.insert_or_assign(std::string{prefix} + "light_color", params.lightColor);
+        rv.insert_or_assign(std::string{prefix} + "background_color", params.backgroundColor);
         // TODO: floorLocation
 
         return rv;
@@ -45,17 +45,17 @@ namespace
 
     void UpdFromValues(
         std::string_view prefix,
-        const std::unordered_map<std::string, AppSettingValue>& values,
+        const std::unordered_map<std::string, Variant>& values,
         ModelRendererParams& params)
     {
         params.decorationOptions.tryUpdFromValues(std::string{prefix} + "decorations/", values);
         params.overlayOptions.tryUpdFromValues(std::string{prefix} + "overlays/", values);
         params.renderingOptions.tryUpdFromValues(std::string{prefix} + "graphics/", values);
         if (const auto* v = lookup_or_nullptr(values, std::string{prefix} + "light_color")) {
-            params.lightColor = v->to_color();
+            params.lightColor = v->to<Color>();
         }
         if (const auto* v = lookup_or_nullptr(values,std::string{prefix} + "background_color")) {
-            params.backgroundColor = v->to_color();
+            params.backgroundColor = v->to<Color>();
         }
         // TODO: floorLocation
     }
@@ -66,8 +66,7 @@ osc::ModelRendererParams::ModelRendererParams() :
     backgroundColor{SceneRendererParams::default_background_color()},
     floorLocation{SceneRendererParams::default_floor_location()},
     camera{create_camera_with_radius(5.0f)}
-{
-}
+{}
 
 void osc::UpdModelRendererParamsFrom(
     const AppSettings& settings,